update: now the teachers can modify the grades by adding a float grade  (ex: 7.3, 8.5 etc)

+++ backend/src/main/java/com/eduquizz/backend/controllers/QuizAttemptController.java
@@ -56,7 +56,7 @@ public class QuizAttemptController {
     public ResponseEntity<?> gradeQuestion(
             @PathVariable Long attemptId,
             @PathVariable Long questionId,
-            @RequestParam Integer points) {
+            @RequestParam Double points) {
         try {
             QuizAttempt updatedAttempt = quizAttemptService.gradeQuizAttempt(attemptId, questionId, points);
             return ResponseEntity.ok(updatedAttempt);
